requirements.txt
Python > 3.7
PyTorch
Transformers
scikit-learn
pandas
numpy
tqdm